requirements.txt
aiohttp
aiosignal
async-timeout
azure-core
azure-cosmos 
azure-functions
azure-storage-blob
beautifulsoup4
isodate
multidict
numpy
pandas
pyodbc
python-dotenv
requests
SQLAlchemy
urllib3
playwright
lxml
html5lib